+++ packages/cli/src/commands/generate/config/generate_config_command.test.ts
@@ -3,10 +3,7 @@ import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
 import { GenerateConfigCommand } from './generate_config_command.js';
 import { ClientConfigFormat } from '@aws-amplify/client-config';
 import yargs, { CommandModule } from 'yargs';
-import {
-  TestCommandError,
-  TestCommandRunner,
-} from '../../../test-utils/command_runner.js';
+import { TestCommandRunner } from '../../../test-utils/command_runner.js';
 import assert from 'node:assert';
 import { BackendIdentifierResolver } from '../../../backend-identifier/backend_identifier_resolver.js';
 import { ClientConfigGeneratorAdapter } from '../../../client-config/client_config_generator_adapter.js';
@@ -82,7 +79,7 @@ void describe('generate config command', () => {
 
   void it('generates and writes config for appID and branch', async () => {
     await commandRunner.runCommand(
-      'config --branch branch_name --app-id app_id --out-dir /foo/bar --format js'
+      'config --branch branch_name --app-id app_id --out-dir /foo/bar --format mjs'
     );
     assert.equal(generateClientConfigMock.mock.callCount(), 1);
     assert.deepEqual(generateClientConfigMock.mock.calls[0].arguments[0], {
@@ -96,7 +93,7 @@ void describe('generate config command', () => {
     );
     assert.deepStrictEqual(
       generateClientConfigMock.mock.calls[0].arguments[2],
-      ClientConfigFormat.JS
+      ClientConfigFormat.MJS
     );
   });
 
@@ -121,7 +118,7 @@ void describe('generate config command', () => {
 
   void it('can generate to custom relative path', async () => {
     await commandRunner.runCommand(
-      'config --stack stack_name --out-dir foo/bar --format js'
+      'config --stack stack_name --out-dir foo/bar --format mjs'
     );
     assert.equal(generateClientConfigMock.mock.callCount(), 1);
     assert.deepEqual(generateClientConfigMock.mock.calls[0].arguments[0], {
@@ -134,7 +131,26 @@ void describe('generate config command', () => {
     );
     assert.equal(
       generateClientConfigMock.mock.calls[0].arguments[2],
-      ClientConfigFormat.JS
+      ClientConfigFormat.MJS
+    );
+  });
+
+  void it('can generate config in dart format', async () => {
+    await commandRunner.runCommand(
+      'config --stack stack_name --out-dir foo/bar --format dart'
+    );
+    assert.equal(generateClientConfigMock.mock.callCount(), 1);
+    assert.deepEqual(generateClientConfigMock.mock.calls[0].arguments[0], {
+      stackName: 'stack_name',
+    });
+    assert.equal(generateClientConfigMock.mock.callCount(), 1);
+    assert.equal(
+      generateClientConfigMock.mock.calls[0].arguments[1],
+      'foo/bar'
+    );
+    assert.equal(
+      generateClientConfigMock.mock.calls[0].arguments[2],
+      ClientConfigFormat.DART
     );
   });
 
@@ -148,14 +164,9 @@ void describe('generate config command', () => {
   });
 
   void it('fails if both stack and branch are present', async () => {
-    await assert.rejects(
-      () => commandRunner.runCommand('config --stack foo --branch baz'),
-      (err: TestCommandError) => {
-        assert.equal(err.error.name, 'YError');
-        assert.match(err.error.message, /Arguments .* mutually exclusive/);
-        assert.match(err.output, /Arguments .* are mutually exclusive/);
-        return true;
-      }
+    const output = await commandRunner.runCommand(
+      'config --stack foo --branch baz'
     );
+    assert.match(output, /Arguments .* mutually exclusive/);
   });
 });