+++ packages/cli/src/commands/sandbox/sandbox-secret/sandbox_secret_set_command.ts
@@ -5,6 +5,7 @@ import { SandboxBackendIdentifier } from '@aws-amplify/platform-core';
 import { AmplifyPrompter } from '@aws-amplify/cli-core';
 
 import { ArgumentsKebabCase } from '../../../kebab_case.js';
+import { handleCommandFailure } from '../../../command_failure_handler.js';
 
 /**
  * Command to set sandbox secret.
@@ -50,11 +51,16 @@ export class SandboxSecretSetCommand
    * @inheritDoc
    */
   builder = (yargs: Argv): Argv<SecretSetCommandOptions> => {
-    return yargs.positional('secret-name', {
-      describe: 'Name of the secret to set',
-      type: 'string',
-      demandOption: true,
-    });
+    return yargs
+      .positional('secret-name', {
+        describe: 'Name of the secret to set',
+        type: 'string',
+        demandOption: true,
+      })
+      .fail((msg, err) => {
+        handleCommandFailure(msg, err, yargs);
+        yargs.exit(1, err);
+      });
   };
 }