+++ packages/cli/src/commands/sandbox/sandbox_command.ts
@@ -11,6 +11,8 @@ import {
   DEFAULT_UI_PATH,
 } from '../../form-generation/default_form_generation_output_paths.js';
 import { ArgumentsKebabCase } from '../../kebab_case.js';
+import { fromIni } from '@aws-sdk/credential-provider-ini';
+import { handleCommandFailure } from '../../command_failure_handler.js';
 
 export type SandboxCommandOptions =
   ArgumentsKebabCase<SandboxCommandOptionsCamelCase>;
@@ -67,7 +69,8 @@ export class SandboxCommand
   constructor(
     private readonly sandboxFactory: SandboxSingletonFactory,
     private readonly sandboxSubCommands: CommandModule[],
-    private readonly sandboxEventHandlerCreator?: SandboxEventHandlerCreator
+    private readonly sandboxEventHandlerCreator?: SandboxEventHandlerCreator,
+    private readonly profileCredentialProvider = fromIni
   ) {
     this.command = 'sandbox';
     this.describe = 'Starts sandbox, watch mode for amplify deployments';
@@ -79,6 +82,8 @@ export class SandboxCommand
   handler = async (args: SandboxCommandOptions): Promise<void> => {
     const { profile } = args;
     if (profile) {
+      // check if we can load the profile first
+      await this.profileCredentialProvider({ profile })();
       process.env.AWS_PROFILE = profile;
     }
     const sandbox = await this.sandboxFactory.getInstance();
@@ -201,6 +206,10 @@ export class SandboxCommand
             }
           }
           return true;
+        })
+        .fail((msg, err) => {
+          handleCommandFailure(msg, err, yargs);
+          yargs.exit(1, err);
         })
     );
   };