+++ packages/create-amplify/src/create_amplify.ts
@@ -14,6 +14,8 @@ import { InitialProjectFileGenerator } from './initial_project_file_generator.js
 import { NpmProjectInitializer } from './npm_project_initializer.js';
 import { TsConfigInitializer } from './tsconfig_initializer.js';
 import { getProjectRoot } from './get_project_root.js';
+import { PackageJsonReader } from './package_json_reader.js';
+import { GitIgnoreInitializer } from './gitignore_initializer.js';
 
 const projectRoot = await getProjectRoot();
 
@@ -22,7 +24,8 @@ const amplifyProjectCreator = new AmplifyProjectCreator(
   new ProjectRootValidator(projectRoot),
   new InitialProjectFileGenerator(projectRoot),
   new NpmProjectInitializer(projectRoot),
-  new TsConfigInitializer(projectRoot),
+  new TsConfigInitializer(projectRoot, new PackageJsonReader(projectRoot)),
+  new GitIgnoreInitializer(projectRoot),
   projectRoot
 );