+++ packages/deployed-backend-client/src/deployed-backend-client/deployed_resources_enumerator.test.ts
@@ -11,10 +11,17 @@ import {
 } from '../deployed_backend_client_factory.js';
 import { DeployedResourcesEnumerator } from './deployed_resources_enumerator.js';
 import { StackStatusMapper } from './stack_status_mapper.js';
+import { ArnGenerator } from './arn_generator.js';
+import { ArnParser } from './arn_parser.js';
 
 void describe('listDeployedResources', () => {
+  const arnGeneratorMock = new ArnGenerator();
+  const arnParserMock = new ArnParser();
+  mock.method(arnGeneratorMock, 'generateArn', () => undefined);
   const deployedResourcesEnumerator = new DeployedResourcesEnumerator(
-    new StackStatusMapper()
+    new StackStatusMapper(),
+    arnGeneratorMock,
+    arnParserMock
   );
   const cfnClientSendMock = mock.fn();
   const mockCfnClient = new CloudFormation();
@@ -132,7 +139,9 @@ void describe('listDeployedResources', () => {
     const deployedResources =
       await deployedResourcesEnumerator.listDeployedResources(
         mockCfnClient,
-        'testRootStack'
+        'testRootStack',
+        '0000000',
+        'us-east-1'
       );
 
     const expectedResources: DeployedBackendResource[] = [
@@ -143,6 +152,7 @@ void describe('listDeployedResources', () => {
         resourceStatusReason: undefined,
         resourceType: 'AWS::AppSync::API',
         physicalResourceId: 'apiSubStackAppSyncPhysicalResourceId',
+        arn: undefined,
       },
       {
         logicalResourceId: 'apiStackUserPoolLogicalResourceId',
@@ -151,6 +161,7 @@ void describe('listDeployedResources', () => {
         resourceStatusReason: undefined,
         resourceType: 'AWS::Cognito::UserPool',
         physicalResourceId: 'apiStackUserPoolPhysicalResourceId',
+        arn: undefined,
       },
       {
         logicalResourceId: 'authStackUserPoolLogicalResourceId',
@@ -159,6 +170,7 @@ void describe('listDeployedResources', () => {
         resourceStatusReason: undefined,
         resourceType: 'AWS::Cognito::UserPool',
         physicalResourceId: 'authStackUserPoolPhysicalResourceId',
+        arn: undefined,
       },
       {
         logicalResourceId: 'rootStackIamRoleLogicalResourceId',
@@ -167,6 +179,7 @@ void describe('listDeployedResources', () => {
         resourceStatusReason: undefined,
         resourceType: 'AWS::IAM::Role',
         physicalResourceId: 'rootStackIamRolePhysicalResourceId',
+        arn: undefined,
       },
     ];
     assert.deepEqual(deployedResources, expectedResources);