+++ packages/deployed-backend-client/src/deployed_backend_client.ts
@@ -31,7 +31,6 @@ import {
 } from '@aws-sdk/client-cloudformation';
 
 import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3';
-
 import {
   authOutputKey,
   graphqlOutputKey,
@@ -40,6 +39,7 @@ import {
 } from '@aws-amplify/backend-output-schemas';
 import { DeployedResourcesEnumerator } from './deployed-backend-client/deployed_resources_enumerator.js';
 import { StackStatusMapper } from './deployed-backend-client/stack_status_mapper.js';
+import { ArnParser } from './deployed-backend-client/arn_parser.js';
 
 /**
  * Deployment Client
@@ -53,7 +53,8 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
     private readonly s3Client: S3Client,
     private readonly backendOutputClient: BackendOutputClient,
     private readonly deployedResourcesEnumerator: DeployedResourcesEnumerator,
-    private readonly stackStatusMapper: StackStatusMapper
+    private readonly stackStatusMapper: StackStatusMapper,
+    private readonly arnParser: ArnParser
   ) {}
 
   /**
@@ -214,6 +215,10 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
       nestedStack?.StackName?.includes('data')
     );
 
+    // stack?.StackId is the ARN of the stack
+    const { accountId, region } = this.arnParser.tryParseArn(
+      stack?.StackId as string
+    );
     const backendMetadataObject: BackendMetadata = {
       deploymentType: backendOutput[stackOutputKey].payload
         .deploymentType as BackendDeploymentType,
@@ -222,7 +227,9 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
       name: stackName,
       resources: await this.deployedResourcesEnumerator.listDeployedResources(
         this.cfnClient,
-        stackName
+        stackName,
+        accountId,
+        region
       ),
     };
 
@@ -248,17 +255,6 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
     }
 
     if (apiStack) {
-      const schemaFileUri =
-        backendOutput[graphqlOutputKey]?.payload.amplifyApiModelSchemaS3Uri;
-      const modelIntrospectionSchemaUri =
-        backendOutput[graphqlOutputKey]?.payload
-          .amplifyApiModelIntrospectionSchemaS3Uri ||
-        // TODO: Once the introspection schema is in the CFN output, remove the below line and the OR operator above.
-        // GH link: https://github.com/aws-amplify/samsara-cli/issues/395
-        `${
-          schemaFileUri.slice(0, schemaFileUri.lastIndexOf('/')) as string
-        }/model-introspection-schema.json`;
-
       const additionalAuthTypesString =
         backendOutput[graphqlOutputKey]?.payload
           .awsAppsyncAdditionalAuthenticationTypes;
@@ -277,12 +273,8 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
         additionalAuthTypes,
         conflictResolutionMode: backendOutput[graphqlOutputKey]?.payload
           .awsAppsyncConflictResolutionMode as ConflictResolutionMode,
-        graphqlSchema: await this.fetchSchema(schemaFileUri),
         apiId: backendOutput[graphqlOutputKey]?.payload
           .awsAppsyncApiId as string,
-        modelIntrospectionSchema: await this.fetchSchema(
-          modelIntrospectionSchemaUri
-        ),
       };
     }