+++ packages/integration-tests/src/create_empty_amplify_project.ts
@@ -20,7 +20,11 @@ export const createEmptyAmplifyProject = async (
   const projectName = `${TEST_PROJECT_PREFIX}-${projectDirName}-${shortUuid()}`;
   await fs.writeFile(
     path.join(projectRoot, 'package.json'),
-    JSON.stringify({ name: projectName }, null, 2)
+    JSON.stringify(
+      { name: projectName, type: 'module' },
+      null,
+      2
+    )
   );
 
   const projectAmplifyDir = path.join(projectRoot, 'amplify');