+++ packages/integration-tests/src/test-e2e/test_project.ts
@@ -3,6 +3,7 @@ import { DataStorageAuthWithTriggerTestProject } from './data_storage_auth_with_
 import { getSecretClient } from '@aws-amplify/backend-secret';
 import { createTestDirectory } from '../setup_test_directory.js';
 import { TestProjectBase } from './test_project_base.js';
+import { MinimalWithTypescriptIdiomTestProject } from './minimalist-project-with-typescript-idioms.js';
 
 /**
  * Generates a list of test projects.
@@ -19,6 +20,10 @@ export const generateTestProjects = async (
       e2eProjectDir,
       cfnClient,
       secretClient
+    ),
+    await MinimalWithTypescriptIdiomTestProject.createProject(
+      e2eProjectDir,
+      cfnClient
     )
   );
   return testProjects;