+++ packages/cli-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/cli-core
 
+## 0.1.0-alpha.4
+
+### Patch Changes
+
+- 76ba7929: chore: reformat yargs error display
+
+## 0.1.0-alpha.3
+
+### Patch Changes
+
+- 675f4283: Move prompter in create_amplify to cli-core package
+
 ## 0.1.0-alpha.2
 
 ### Minor Changes